Store a direct reference to the sub list instead of a binsha.
This avoids many dict lookups and makes the data structure
significantly smaller too. For a large kernel tree it shrank
from 20GB in memory to 3.5GB in memory. The pickle file only
went from 1.8GB to 1.7GB though, indicating that this has
eliminated a large amount of python object overhead.